Add error message popup and input validation to all class that needs it.

+++ Incharge/Service/ClientService.cs
@@ -50,10 +50,10 @@ namespace Incharge.Service
         }
         public void AddService(ClientVM clientVM) //cosider if there is a desire to add membership as soon as client profile is made..
         {
-            if(clientVM == null) { throw new NullReferenceException("Input Empty."); }
+            if(clientVM == null || clientVM.FirstName == null || clientVM.LastName == null) { throw new NullReferenceException("Input Empty."); }
             var result = _photoService.AddPhotoAsync(clientVM.PicutreInput).Result;
             //Make whole program async when all tests are completed
-            if(result == null) { throw new NullReferenceException("Photo Empty or Invalid."); }
+            if(result == null || result.Url==null) { throw new NullReferenceException("Photo Empty or Invalid."); }
             var client = _mapper.Map<Client>(clientVM); //some is null so check if this is true
             client.ProfilePicture = result.Url.ToString();
             _ClientRepository.Add(client) ;
@@ -62,17 +62,19 @@ namespace Incharge.Service
         public void UpdateService(ClientVM clientVM) //email, phone, firstname and lastname are required
         {
             var clientToUpdate = _FindClientRepository.FindBy(x => x.Uuid == clientVM.Uuid);
-            if(clientToUpdate.ProfilePicture != null)
+            if (clientVM.PicutreInput != null)
             {
-				var delete = _photoService.DeletePhotoAsync(clientToUpdate.ProfilePicture).Result;
-				if (clientToUpdate == null) { throw new NullReferenceException("Client Empty."); }
-                Console.WriteLine(delete.ToString());
-			}
-            if(clientVM.PicutreInput != null) 
-            {
-				var result = _photoService.AddPhotoAsync(clientVM.PicutreInput).Result;
-				clientVM.ProfilePicture = result.Url.ToString();
-			}
+                if (clientToUpdate.ProfilePicture != null && clientVM.PicutreInput != null)
+                {
+                    var delete = _photoService.DeletePhotoAsync(clientToUpdate.ProfilePicture).Result;
+                    if (clientToUpdate == null) { throw new NullReferenceException("Client Empty."); }
+                    Console.WriteLine(delete.ToString());
+                }
+
+                var result = _photoService.AddPhotoAsync(clientVM.PicutreInput).Result;
+                clientVM.ProfilePicture = result.Url.ToString();
+            }
+            
 			var phone = clientToUpdate.Phone;
 			_mapper.Map(clientVM, clientToUpdate);
             if(clientToUpdate.Phone == 0) { clientToUpdate.Phone = phone; }
